feat: tmp

+++ genome.go
@@ -2,13 +2,13 @@ package goga
 
 // IGenome associates a fitness with a bitset
 type Genome interface {
-	GetFitness() int
-	SetFitness(int)
+	GetFitness() float64
+	SetFitness(float64)
 	GetBits() *Bitset
 }
 
 type genome struct {
-	fitness int
+	fitness float64
 	bitset  Bitset
 }
 
@@ -18,11 +18,11 @@ func NewGenome(bitset Bitset) Genome {
 	return &genome{bitset: bitset}
 }
 
-func (g *genome) GetFitness() int {
+func (g *genome) GetFitness() float64 {
 	return g.fitness
 }
 
-func (g *genome) SetFitness(fitness int) {
+func (g *genome) SetFitness(fitness float64) {
 	g.fitness = fitness
 }
